+++ 17-04/17-04.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33516.290
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "17-04", "17-04\17-04.csproj", "{44C5C9AD-9419-49F0-991E-BDD823FE9CB1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{44C5C9AD-9419-49F0-991E-BDD823FE9CB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44C5C9AD-9419-49F0-991E-BDD823FE9CB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44C5C9AD-9419-49F0-991E-BDD823FE9CB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44C5C9AD-9419-49F0-991E-BDD823FE9CB1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E9A37BF9-C36D-4225-BCC6-43BD75F7849B}
+	EndGlobalSection
+EndGlobal
